requirements.txt
absl-py==0.4.1
argh==0.26.2
astor==0.7.1
astroid==2.0.4
beautifulsoup4==4.6.3
bleach==1.5.0
certifi==2018.8.24
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
gast==0.2.0
grpcio==1.14.2
html5lib==0.9999999
idna==2.7
isort==4.3.4
jieba3k==0.35.1
jsonrpclib-pelix==0.3.1
lazy-object-proxy==1.3.1
lxml==4.2.4
Markdown==2.6.11
mccabe==0.6.1
newspaper3k==0.2.6
nltk==3.3
numpy==1.14.5
pandas==0.23.4
pathtools==0.1.2
pika==0.12.0
Pillow==5.2.0
protobuf==3.6.1
pylint==2.1.1
pymongo==3.7.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-file==1.4.3
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
tldextract==2.2.0
typed-ast==1.1.0
urllib3==1.23
watchdog==0.9.0
Werkzeug==0.14.1
wrapt==1.10.11
